Move rejoin_tokens to verus_syn
