+++ dependencies/syn/src/lib.rs
@@ -1024,3 +1024,124 @@ pub fn parse_file(mut content: &str) -> Result<File> {
     file.shebang = shebang;
     Ok(file)
 }
+
+pub fn rejoin_tokens(stream: proc_macro2::TokenStream) -> proc_macro2::TokenStream {
+    use proc_macro2::{Group, Punct, Spacing::*, Span, TokenTree};
+    let mut tokens: Vec<TokenTree> = stream.into_iter().collect();
+    let pun = |t: &TokenTree| match t {
+        TokenTree::Punct(p) => Some((p.as_char(), p.spacing(), p.span())),
+        _ => None,
+    };
+    let ident = |t: &TokenTree| match t {
+        TokenTree::Ident(p) => Some((p.to_string(), p.span())),
+        _ => None,
+    };
+    let adjacent = |s1: Span, s2: Span| {
+        let l1 = s1.end();
+        let l2 = s2.start();
+        l1.eq(&l2)
+    };
+    fn mk_joint_punct(t: Option<(char, proc_macro2::Spacing, Span)>) -> TokenTree {
+        let (op, _, span) = t.unwrap();
+        let mut punct = Punct::new(op, Joint);
+        punct.set_span(span);
+        TokenTree::Punct(punct)
+    }
+    let mut i = 0;
+    let mut till = if tokens.len() >= 2 {
+        tokens.len() - 2
+    } else {
+        0
+    };
+    while i < till {
+        let t0 = pun(&tokens[i]);
+        let t1_ident = ident(&tokens[i + 1]);
+        match (t0, t1_ident.as_ref().map(|(a, b)| (a.as_str(), *b))) {
+            (Some(('!', Alone, s1)), Some(("is", s2))) => {
+                if adjacent(s1, s2) {
+                    tokens[i] =
+                        TokenTree::Ident(proc_macro2::Ident::new("isnt", s1.join(s2).unwrap()));
+                    tokens.remove(i + 1);
+                    i += 1;
+                    till -= 1;
+                    continue;
+                }
+            }
+            (Some(('!', Alone, s1)), Some(("has", s2))) => {
+                if adjacent(s1, s2) {
+                    tokens[i] =
+                        TokenTree::Ident(proc_macro2::Ident::new("hasnt", s1.join(s2).unwrap()));
+                    tokens.remove(i + 1);
+                    i += 1;
+                    till -= 1;
+                    continue;
+                }
+            }
+            _ => {}
+        }
+        let t1_pun = pun(&tokens[i + 1]);
+        let t2 = pun(&tokens[i + 2]);
+        let t3 = if i + 3 < tokens.len() {
+            pun(&tokens[i + 3])
+        } else {
+            None
+        };
+        match (t0, t1_pun, t2, t3) {
+            (
+                Some(('<', Joint, _)),
+                Some(('=', Alone, s1)),
+                Some(('=', Joint, s2)),
+                Some(('>', Alone, _)),
+            )
+            | (Some(('=', Joint, _)), Some(('=', Alone, s1)), Some(('=', Alone, s2)), _)
+            | (Some(('!', Joint, _)), Some(('=', Alone, s1)), Some(('=', Alone, s2)), _)
+            | (Some(('=', Joint, _)), Some(('=', Alone, s1)), Some(('>', Alone, s2)), _)
+            | (Some(('<', Joint, _)), Some(('=', Alone, s1)), Some(('=', Alone, s2)), _)
+            | (Some(('&', Joint, _)), Some(('&', Alone, s1)), Some(('&', Alone, s2)), _)
+            | (Some(('|', Joint, _)), Some(('|', Alone, s1)), Some(('|', Alone, s2)), _) => {
+                if adjacent(s1, s2) {
+                    tokens[i + 1] = mk_joint_punct(t1_pun);
+                }
+            }
+            (Some(('=', Alone, _)), Some(('~', Alone, s1)), Some(('=', Alone, s2)), _)
+            | (Some(('!', Alone, _)), Some(('~', Alone, s1)), Some(('=', Alone, s2)), _) => {
+                if adjacent(s1, s2) {
+                    tokens[i] = mk_joint_punct(t0);
+                    tokens[i + 1] = mk_joint_punct(t1_pun);
+                }
+            }
+            (
+                Some(('=', Alone, _)),
+                Some(('~', Alone, _)),
+                Some(('~', Alone, s2)),
+                Some(('=', Alone, s3)),
+            )
+            | (
+                Some(('!', Alone, _)),
+                Some(('~', Alone, _)),
+                Some(('~', Alone, s2)),
+                Some(('=', Alone, s3)),
+            ) => {
+                if adjacent(s2, s3) {
+                    tokens[i] = mk_joint_punct(t0);
+                    tokens[i + 1] = mk_joint_punct(t1_pun);
+                    tokens[i + 2] = mk_joint_punct(t2);
+                }
+            }
+            _ => {}
+        }
+        i += 1;
+    }
+    for tt in &mut tokens {
+        match tt {
+            TokenTree::Group(group) => {
+                let mut new_group = Group::new(group.delimiter(), rejoin_tokens(group.stream()));
+                new_group.set_span(group.span());
+                *group = new_group;
+            }
+            _ => {}
+        }
+    }
+    use std::iter::FromIterator;
+    proc_macro2::TokenStream::from_iter(tokens.into_iter())
+}
